Add popup functionality
